+++ frontend/src/app/features/admin/components/admin-event-display/admin-event-display.component.spec.ts
@@ -5,8 +5,8 @@ import {EventsService} from "../../../../core/services/events.service";
 import {MatDialog, MatDialogRef} from "@angular/material/dialog";
 import {NO_ERRORS_SCHEMA, Provider} from "@angular/core";
 import {EventDto} from "../../../../core/api";
-import {EventEditorDialogComponent} from "../event-editor/event-editor-dialog.component";
-import {EventEditorMode} from "../../enums/event-editor-mode";
+import {EventEditorDialogComponent} from "../event-editor-dialog/event-editor-dialog.component";
+import {EditorMode} from "../../enums/editor-mode";
 import {EventEditorData} from "../../models/event-editor-data";
 import {of} from "rxjs";
 
@@ -19,7 +19,7 @@ describe('AdminEventDisplayComponent', () => {
   let matDialogSpy: jasmine.SpyObj<MatDialog>;
 
   beforeEach(waitForAsync(() => {
-    stateSpy = jasmine.createSpyObj<ApplicationStateService>([], ['events'])
+    stateSpy = jasmine.createSpyObj<ApplicationStateService>([], ['events$'])
     eventsServiceSpy = jasmine.createSpyObj<EventsService>(['getEvents', 'createEvent', 'updateEvent', 'deleteEvent'])
     matDialogSpy = jasmine.createSpyObj<MatDialog>(['open'])
 
@@ -76,7 +76,7 @@ describe('AdminEventDisplayComponent', () => {
     expect(matDialogSpy.open).toHaveBeenCalledOnceWith(EventEditorDialogComponent, {
       data: {
         event: emptyEventWithDates,
-        mode: EventEditorMode.Create
+        mode: EditorMode.Create
       } as EventEditorData
     });
     expect(eventsServiceSpy.createEvent).toHaveBeenCalledOnceWith(createdEvent);
@@ -124,7 +124,7 @@ describe('AdminEventDisplayComponent', () => {
     expect(matDialogSpy.open).toHaveBeenCalledWith(EventEditorDialogComponent, {
       data: {
         event: eventToUpdate,
-        mode: EventEditorMode.Update
+        mode: EditorMode.Update
       } as EventEditorData
     });
     expect(eventsServiceSpy.updateEvent).toHaveBeenCalledWith(updatedEvent);
@@ -155,7 +155,7 @@ describe('AdminEventDisplayComponent', () => {
     expect(matDialogSpy.open).toHaveBeenCalledWith(EventEditorDialogComponent, {
       data: {
         event: eventToUpdate,
-        mode: EventEditorMode.Update
+        mode: EditorMode.Update
       } as EventEditorData
     });
     expect(eventsServiceSpy.deleteEvent).toHaveBeenCalledWith(eventId);
@@ -187,7 +187,7 @@ describe('AdminEventDisplayComponent', () => {
     expect(matDialogSpy.open).toHaveBeenCalledWith(EventEditorDialogComponent, {
       data: {
         event: expectedEvent,
-        mode: EventEditorMode.Create
+        mode: EditorMode.Create
       } as EventEditorData
     });
     expect(eventsServiceSpy.createEvent).not.toHaveBeenCalled();
